+++ contracts/src/v0.7/ChainlinkClient.sol
@@ -4,6 +4,7 @@ pragma solidity ^0.7.0;
 import "./Chainlink.sol";
 import "./interfaces/ENSInterface.sol";
 import "./interfaces/LinkTokenInterface.sol";
+import "./interfaces/ChainlinkRequestInterface.sol";
 import "./interfaces/OperatorInterface.sol";
 import "./interfaces/PointerInterface.sol";
 import { ENSResolver as ENSResolver_Chainlink } from "./vendor/ENSResolver.sol";
@@ -13,7 +14,7 @@ import { ENSResolver as ENSResolver_Chainlink } from "./vendor/ENSResolver.sol";
  * @notice Contract writers can inherit this contract in order to create requests for the
  * Chainlink network
  */
-contract ChainlinkClient {
+abstract contract ChainlinkClient {
   using Chainlink for Chainlink.Request;
 
   uint256 constant internal LINK_DIVISIBILITY = 10**18;
@@ -25,12 +26,12 @@ contract ChainlinkClient {
   bytes32 constant private ENS_ORACLE_SUBNAME = keccak256("oracle");
   address constant private LINK_TOKEN_POINTER = 0xC89bD4E1632D3A43CB03AAAd5262cbe4038Bc571;
 
-  ENSInterface private ens;
-  bytes32 private ensNode;
-  LinkTokenInterface private link;
-  OperatorInterface private oracle;
-  uint256 private requestCount = 1;
-  mapping(bytes32 => address) private pendingRequests;
+  ENSInterface private s_ens;
+  bytes32 private s_ensNode;
+  LinkTokenInterface private s_link;
+  OperatorInterface private s_oracle;
+  uint256 private s_requestCount = 1;
+  mapping(bytes32 => address) private s_pendingRequests;
 
   event ChainlinkRequested(
     bytes32 indexed id
@@ -45,13 +46,13 @@ contract ChainlinkClient {
   /**
    * @notice Creates a request that can hold additional parameters
    * @param specId The Job Specification ID that the request will be created for
-   * @param callbackAddress The callback address that the response will be sent to
-   * @param callbackFunctionSignature The callback function signature to use for the callback address
+   * @param callbackAddr address to operate the callback on
+   * @param callbackFunctionSignature function signature to use for the callback
    * @return A Chainlink Request struct in memory
    */
   function buildChainlinkRequest(
     bytes32 specId,
-    address callbackAddress,
+    address callbackAddr,
     bytes4 callbackFunctionSignature
   )
     internal
@@ -61,7 +62,27 @@ contract ChainlinkClient {
     )
   {
     Chainlink.Request memory req;
-    return req.initialize(specId, callbackAddress, callbackFunctionSignature);
+    return req.initialize(specId, callbackAddr, callbackFunctionSignature);
+  }
+
+  /**
+   * @notice Creates a request that can hold additional parameters
+   * @param specId The Job Specification ID that the request will be created for
+   * @param callbackFunctionSignature function signature to use for the callback
+   * @return A Chainlink Request struct in memory
+   */
+  function buildOperatorRequest(
+    bytes32 specId,
+    bytes4 callbackFunctionSignature
+  )
+    internal
+    view
+    returns (
+      Chainlink.Request memory
+    )
+  {
+    Chainlink.Request memory req;
+    return req.initialize(specId, address(this), callbackFunctionSignature);
   }
 
   /**
@@ -80,7 +101,7 @@ contract ChainlinkClient {
       bytes32
     )
   {
-    return sendChainlinkRequestTo(address(oracle), req, payment);
+    return sendChainlinkRequestTo(address(s_oracle), req, payment);
   }
 
   /**
@@ -103,18 +124,30 @@ contract ChainlinkClient {
       bytes32 requestId
     )
   {
-    return rawRequest(oracleAddress, req, payment, ORACLE_ARGS_VERSION, oracle.oracleRequest.selector);
+    uint256 nonce = s_requestCount;
+    s_requestCount = nonce + 1;
+    bytes memory encodedRequest = abi.encodeWithSelector(
+      ChainlinkRequestInterface.oracleRequest.selector,
+      SENDER_OVERRIDE, // Sender value - overridden by onTokenTransfer by the requesting contract's address
+      AMOUNT_OVERRIDE, // Amount value - overridden by onTokenTransfer by the actual amount of LINK sent
+      req.id,
+      address(this),
+      req.callbackFunctionId,
+      nonce,
+      ORACLE_ARGS_VERSION,
+      req.buf.buf);
+    return _rawRequest(oracleAddress, nonce, payment, encodedRequest);
   }
 
   /**
    * @notice Creates a Chainlink request to the stored oracle address
    * @dev This function supports multi-word response
-   * @dev Calls `requestOracleDataFrom` with the stored oracle address
+   * @dev Calls `sendOperatorRequestTo` with the stored oracle address
    * @param req The initialized Chainlink Request
    * @param payment The amount of LINK to send for the request
    * @return requestId The request ID
    */
-  function requestOracleData(
+  function sendOperatorRequest(
     Chainlink.Request memory req,
     uint256 payment
   )
@@ -123,7 +156,7 @@ contract ChainlinkClient {
       bytes32
     )
   {
-    return requestOracleDataFrom(address(oracle), req, payment);
+    return sendOperatorRequestTo(address(s_oracle), req, payment);
   }
 
   /**
@@ -137,7 +170,7 @@ contract ChainlinkClient {
    * @param payment The amount of LINK to send for the request
    * @return requestId The request ID
    */
-  function requestOracleDataFrom(
+  function sendOperatorRequestTo(
     address oracleAddress,
     Chainlink.Request memory req,
     uint256 payment
@@ -147,45 +180,43 @@ contract ChainlinkClient {
       bytes32 requestId
     )
   {
-    return rawRequest(oracleAddress, req, payment, OPERATOR_ARGS_VERSION, oracle.requestOracleData.selector);
+    uint256 nonce = s_requestCount;
+    s_requestCount = nonce + 1;
+    bytes memory encodedRequest = abi.encodeWithSelector(
+      OperatorInterface.operatorRequest.selector,
+      SENDER_OVERRIDE, // Sender value - overridden by onTokenTransfer by the requesting contract's address
+      AMOUNT_OVERRIDE, // Amount value - overridden by onTokenTransfer by the actual amount of LINK sent
+      req.id,
+      req.callbackFunctionId,
+      nonce,
+      OPERATOR_ARGS_VERSION,
+      req.buf.buf);
+    return _rawRequest(oracleAddress, nonce, payment, encodedRequest);
   }
 
   /**
    * @notice Make a request to an oracle
    * @param oracleAddress The address of the oracle for the request
-   * @param req The initialized Chainlink Request
+   * @param nonce used to generate the request ID
    * @param payment The amount of LINK to send for the request
-   * @param argsVersion The version of data support (single word, multi word)
+   * @param encodedRequest data encoded for request type specific format
    * @return requestId The request ID
    */
-  function rawRequest(
+  function _rawRequest(
     address oracleAddress,
-    Chainlink.Request memory req,
+    uint256 nonce,
     uint256 payment,
-    uint256 argsVersion,
-    bytes4 funcSelector
+    bytes memory encodedRequest
   )
     private
     returns (
       bytes32 requestId
     )
   {
-    requestId = keccak256(abi.encodePacked(this, requestCount));
-    req.nonce = requestCount;
-    pendingRequests[requestId] = oracleAddress;
+    requestId = keccak256(abi.encodePacked(this, nonce));
+    s_pendingRequests[requestId] = oracleAddress;
     emit ChainlinkRequested(requestId);
-    bytes memory encodedData = abi.encodeWithSelector(
-      funcSelector,
-      SENDER_OVERRIDE, // Sender value - overridden by onTokenTransfer by the requesting contract's address
-      AMOUNT_OVERRIDE, // Amount value - overridden by onTokenTransfer by the actual amount of LINK sent
-      req.id,
-      req.callbackAddress,
-      req.callbackFunctionId,
-      req.nonce,
-      argsVersion,
-      req.buf.buf);
-    require(link.transferAndCall(oracleAddress, payment, encodedData), "unable to transferAndCall to oracle");
-    requestCount += 1;
+    require(s_link.transferAndCall(oracleAddress, payment, encodedRequest), "unable to transferAndCall to oracle");
   }
 
   /**
@@ -206,11 +237,27 @@ contract ChainlinkClient {
   )
     internal
   {
-    OperatorInterface requested = OperatorInterface(pendingRequests[requestId]);
-    delete pendingRequests[requestId];
+    OperatorInterface requested = OperatorInterface(s_pendingRequests[requestId]);
+    delete s_pendingRequests[requestId];
     emit ChainlinkCancelled(requestId);
     requested.cancelOracleRequest(requestId, payment, callbackFunc, expiration);
   }
+
+  /**
+   * @notice the next request count to be used in generating a nonce
+   * @dev starts at 1 in order to ensure consistent gas cost
+   * @return returns the next request count to be used in a nonce
+   */
+  function getNextRequestCount()
+    internal
+    view
+    returns (
+      uint256
+    )
+  {
+    return s_requestCount;
+  }
+
 
   /**
    * @notice Sets the stored oracle address
@@ -221,7 +268,7 @@ contract ChainlinkClient {
   )
     internal
   {
-    oracle = OperatorInterface(oracleAddress);
+    s_oracle = OperatorInterface(oracleAddress);
   }
 
   /**
@@ -233,7 +280,7 @@ contract ChainlinkClient {
   )
     internal
   {
-    link = LinkTokenInterface(linkAddress);
+    s_link = LinkTokenInterface(linkAddress);
   }
 
   /**
@@ -257,7 +304,7 @@ contract ChainlinkClient {
       address
     )
   {
-    return address(link);
+    return address(s_link);
   }
 
   /**
@@ -271,7 +318,7 @@ contract ChainlinkClient {
       address
     )
   {
-    return address(oracle);
+    return address(s_oracle);
   }
 
   /**
@@ -287,7 +334,7 @@ contract ChainlinkClient {
     internal
     notPendingRequest(requestId)
   {
-    pendingRequests[requestId] = oracleAddress;
+    s_pendingRequests[requestId] = oracleAddress;
   }
 
   /**
@@ -302,10 +349,10 @@ contract ChainlinkClient {
   )
     internal
   {
-    ens = ENSInterface(ensAddress);
-    ensNode = node;
-    bytes32 linkSubnode = keccak256(abi.encodePacked(ensNode, ENS_TOKEN_SUBNAME));
-    ENSResolver_Chainlink resolver = ENSResolver_Chainlink(ens.resolver(linkSubnode));
+    s_ens = ENSInterface(ensAddress);
+    s_ensNode = node;
+    bytes32 linkSubnode = keccak256(abi.encodePacked(s_ensNode, ENS_TOKEN_SUBNAME));
+    ENSResolver_Chainlink resolver = ENSResolver_Chainlink(s_ens.resolver(linkSubnode));
     setChainlinkToken(resolver.addr(linkSubnode));
     updateChainlinkOracleWithENS();
   }
@@ -317,8 +364,8 @@ contract ChainlinkClient {
   function updateChainlinkOracleWithENS()
     internal
   {
-    bytes32 oracleSubnode = keccak256(abi.encodePacked(ensNode, ENS_ORACLE_SUBNAME));
-    ENSResolver_Chainlink resolver = ENSResolver_Chainlink(ens.resolver(oracleSubnode));
+    bytes32 oracleSubnode = keccak256(abi.encodePacked(s_ensNode, ENS_ORACLE_SUBNAME));
+    ENSResolver_Chainlink resolver = ENSResolver_Chainlink(s_ens.resolver(oracleSubnode));
     setChainlinkOracle(resolver.addr(oracleSubnode));
   }
 
@@ -344,9 +391,9 @@ contract ChainlinkClient {
     bytes32 requestId
   )
   {
-    require(msg.sender == pendingRequests[requestId],
+    require(msg.sender == s_pendingRequests[requestId],
             "Source must be the oracle of the request");
-    delete pendingRequests[requestId];
+    delete s_pendingRequests[requestId];
     emit ChainlinkFulfilled(requestId);
     _;
   }
@@ -359,7 +406,7 @@ contract ChainlinkClient {
     bytes32 requestId
   )
   {
-    require(pendingRequests[requestId] == address(0), "Request is already pending");
+    require(s_pendingRequests[requestId] == address(0), "Request is already pending");
     _;
   }
 }